[lib] Use ‘size_t’ for ‘svz_interface_t.index’.
* src/libserveez/interface.h (svz_interface_t index):
Convert type to ‘size_t’.
(svz_interface_add): Convert 1st arg type to ‘size_t’.
* src/libserveez/interface.c (svz_interface_collect):
Use ‘size_t’ for local var.
(svz_interface_add): Convert 1st arg type to ‘size_t’.

* src/option.c (display_ifc): Use ‘%zu’ in format string.

+++ src/libserveez/interface.c
@@ -522,7 +522,7 @@ svz_interface_collect (void)
       ifr2.ifr_addr.sa_family = AF_INET;
       if (ioctl (fd, SIOCGIFADDR, &ifr2) == 0)
         {
-          static int index = 0;
+          static size_t index = 0;
 
           /*
            * The following cast looks bogus.  ifr2.ifr_addr is a
@@ -579,7 +579,7 @@ destroy_ifc (void *p)
  * the given network interface has been detected by Serveez itself or not.
  */
 int
-svz_interface_add (int index, char *desc, in_addr_t addr, int detected)
+svz_interface_add (size_t index, char *desc, in_addr_t addr, int detected)
 {
   char *p;
   size_t n;
